LayoutTests/editing/execCommand/overtype-support.html

<!DOCTYPE html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('This test checks that the Overwrite mode is: 1) not exposed to JavaScript 2) queryCommandState returns false.');

shouldBeFalse("document.queryCommandSupported('OverWrite')");
shouldBeFalse("document.queryCommandState('OverWrite')");

var successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
